Fix inconsistent behavior with rspec-core in nested describe blocks on helper methods with clashing let methods
If you had an example group that described a module, and contained a let() with a name that clashed with a method in the module, and you had a sub-example group, RSpec and ParallelRSpec would behave slightly differently.

The root of the issue is that RSpec itself behaves differently if the let() is defined in that top group or the sub-group; only the latter works properly, due to a bug they have with clearing conflicting methods.

Now we trigger the same helper inclusion order as rspec-core would, giving us consistent behavior.