Update the rspec persisted file after concluding a run
The rspec persistence file (if configured) tracks the status of the last
run of a given set of specs, and allows you to use features like
"--only-failures". If prspec updates this file then repeated runs of
prspec with --only-failures can be useful for gradually fixing a large
chunk of failures without re-running an entire test suite.